Results stored in sqlite3 databases
The script now makes us of sqlite3 to store results on the go. If the
queries stoppoed, for instance if google blocks the IP, it is possible
to restart from where it stopped.
Time is now mapped throug a timeMapper function. This is not completely
satisfactory as we should have one mapper function for all the tbs
arguments of an url and one mapper for the tbm argument (the corpus
argument).